requirements.txt
python-dotenv==1.0.0
supabase==1.0.3
flask==2.2.3